+++ src/main/java/de/schaack/ml/basics/functions/activation/interfaces/ActivationFunction.java
@@ -1,9 +1,7 @@
 package de.schaack.ml.basics.functions.activation.interfaces;
 
-
 public interface ActivationFunction {
    double activate(double value);
 
-   double derivative(double activatedValue);
+   double deriveActivation(double activatedValue);
 }
-