+++ src/main/java/de/schaack/ml/basics/functions/distribution/GaussianDistribution.java
@@ -1,12 +1,21 @@
 package de.schaack.ml.basics.functions.distribution;
 
+
 //TODO
 public class GaussianDistribution {
-    private double mean = 0;
-    private double variance = 0;
+    private final double mean;
+    private final double variance;
 
     public GaussianDistribution(double mean, double variance){
         this.mean = mean;
         this.variance = variance;
+    }
+
+    public double draw(){
+        return 0.0;
+    }
+
+    public double[] draw(int amountOfDraws){
+        return new double[0];
     }
 }