+++ src/main/java/de/schaack/ml/basics/functions/evaluation/interfaces/EvaluationFunction.java
@@ -2,5 +2,6 @@ package de.schaack.ml.basics.functions.evaluation.interfaces;
 
 public interface EvaluationFunction {
 
+    void evaluate();
     
 }