+++ src/main/java/de/schaack/ml/basics/functions/loss/interfaces/LossFunction.java
@@ -1,7 +1,9 @@
 package de.schaack.ml.basics.functions.loss.interfaces;
 
 public interface LossFunction {
-    double derivativeLoss(double trueLabel, double predictedLabel);
+    double deriveLoss(double trueLabel, double predictedLabel);
+
     double calculateLoss(double trueLabels, double predictedLabels);
+
     double calculateLoss(double[] trueLabels, double[] predictedLabels);
 }