+++ src/main/java/de/schaack/ml/basics/functions/optimizer/implementation/StochasticGradientDescent.java
@@ -1,11 +1,8 @@
 package de.schaack.ml.basics.functions.optimizer.implementation;
 
-import java.util.Arrays;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import de.schaack.ml.basics.functions.math.MatrixOperations;
 import de.schaack.ml.basics.functions.optimizer.interfaces.OptimiserFunction;
 import de.schaack.ml.basics.settings.optimiser.implementation.LearningRateOptimiser;
 
@@ -38,20 +35,12 @@ public class StochasticGradientDescent implements OptimiserFunction {
 
     // w_t+1 = w_t - learning rate * 1/B * Derivative(Sum_b(Loss_b(w_t)))
     @Override
-    public double[] calculateUpdates(double[] valuesOld, double gradient) {
-        double update = settings.getLearningRate() * gradient;
-        double[] newValues = MatrixOperations.valuesSubValue(valuesOld, update);
-        log.debug("The old values {} recieves the update {} based on the gradient {}, resulting in the new value {}",
-                Arrays.toString(valuesOld), gradient, update, Arrays.toString(newValues));
-        return newValues;
-    }
-
-    @Override
-    public double calculateUpdate(double valueOld, double gradient) {
-        double update = settings.getLearningRate() * gradient;
+    public double calculateUpdate(double valueOld, int batchSize, double gradient) {
+        double update = settings.getLearningRate() * (1.0 / batchSize) * gradient;
         double newValue = valueOld - update;
-        log.debug("The old value {} recieves the update {} based on the gradient {}, resulting in the new value {}",
-                valueOld, gradient, update, newValue);
+        log.debug(
+                "The old value {} recieves the update {} based on the gradient {} from a batch of size {}, resulting in the new value {}",
+                valueOld, gradient, batchSize, update, newValue);
         return newValue;
     }