+++ src/main/java/de/schaack/ml/basics/functions/optimizer/interfaces/OptimiserFunction.java
@@ -4,7 +4,5 @@ import de.schaack.ml.basics.settings.optimiser.interfaces.OptimiserSettings;
 
 public interface OptimiserFunction {
     OptimiserSettings getSettings();
-
-    double[] calculateUpdates(double[] valuesOld, double gradient);
-    double calculateUpdate(double valueOld, double gradient);
+    double calculateUpdate(double valueOld, int batchSize, double gradient);
 }