+++ src/main/java/de/schaack/ml/basics/settings/model/implementation/classification/PerceptronModelSettings.java
@@ -2,6 +2,8 @@ package de.schaack.ml.basics.settings.model.implementation.classification;
 
 import de.schaack.ml.basics.functions.activation.implementation.SigmoidActivation;
 import de.schaack.ml.basics.functions.activation.interfaces.ActivationFunction;
+import de.schaack.ml.basics.functions.optimizer.implementation.StochasticGradientDescent;
+import de.schaack.ml.basics.functions.optimizer.interfaces.OptimiserFunction;
 import de.schaack.ml.basics.settings.model.interfaces.ModelSettings;
 
 /**
@@ -11,6 +13,7 @@ public class PerceptronModelSettings implements ModelSettings {
 
     private int numberOfEpochs = 1;
     private ActivationFunction activationFunction = new SigmoidActivation();
+    private OptimiserFunction optimiserFunction = new StochasticGradientDescent(0.1);
 
     public static final String PARAM_LEARNING_RATE = "learningRate";
     public static final String PARAM_NUMBER_OF_EPOCHS = "numberOfEpochs";
@@ -30,6 +33,16 @@ public class PerceptronModelSettings implements ModelSettings {
     @Override
     public ActivationFunction getActivationFunction() {
         return this.activationFunction;
+    }
+
+    /**
+     * 
+     * Retrieves the optimiser function for the model.
+     *
+     * @return the optimiser function.
+     */
+    public OptimiserFunction getOptimiserFunction() {
+        return this.optimiserFunction;
     }
 
     /**